use atomic instead of mutex when counting

+++ app_test.go
@@ -270,7 +270,7 @@ func Test_App_Mount(t *testing.T) {
 	resp, err := app.Test(httptest.NewRequest(MethodGet, "/john/doe", nil))
 	utils.AssertEqual(t, nil, err, "app.Test(req)")
 	utils.AssertEqual(t, 200, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, 2, app.handlerCount)
+	utils.AssertEqual(t, uint32(2), app.handlerCount)
 }
 
 func Test_App_Use_Params(t *testing.T) {
@@ -845,7 +845,7 @@ func Test_App_Group_Mount(t *testing.T) {
 	resp, err := app.Test(httptest.NewRequest(MethodGet, "/v1/john/doe", nil))
 	utils.AssertEqual(t, nil, err, "app.Test(req)")
 	utils.AssertEqual(t, 200, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, 2, app.handlerCount)
+	utils.AssertEqual(t, uint32(2), app.handlerCount)
 }
 
 func Test_App_Group(t *testing.T) {
